+++ src/components/datagrid/data_grid_header_row.tsx
@@ -27,7 +27,10 @@ import {
   EuiDataGridControlColumn,
 } from './data_grid_types';
 import { CommonProps } from '../common';
-import { EuiDataGridSchema } from './data_grid_schema';
+import {
+  EuiDataGridSchema,
+  EuiDataGridSchemaDetector,
+} from './data_grid_schema';
 import { EuiDataGridDataRowProps } from './data_grid_data_row';
 import { EuiDataGridHeaderCell } from './data_grid_header_cell';
 import { EuiDataGridControlHeaderCell } from './data_grid_control_header_cell';
@@ -38,6 +41,7 @@ export interface EuiDataGridHeaderRowPropsSpecificProps {
   columns: EuiDataGridColumn[];
   columnWidths: EuiDataGridColumnWidths;
   schema: EuiDataGridSchema;
+  schemaDetectors: EuiDataGridSchemaDetector[];
   defaultColumnWidth?: number | null;
   setColumnWidth: (columnId: string, width: number) => void;
   setVisibleColumns: (columnId: string[]) => void;
@@ -61,6 +65,7 @@ const EuiDataGridHeaderRow = forwardRef<
     trailingControlColumns = [],
     columns,
     schema,
+    schemaDetectors,
     columnWidths,
     defaultColumnWidth,
     className,
@@ -106,6 +111,7 @@ const EuiDataGridHeaderRow = forwardRef<
           focusedCell={focusedCell}
           onCellFocus={setFocusedCell}
           schema={schema}
+          schemaDetectors={schemaDetectors}
           setColumnWidth={setColumnWidth}
           setVisibleColumns={setVisibleColumns}
           switchColumnPos={switchColumnPos}