+++ src/components/image/image.tsx
@@ -38,6 +38,8 @@ import { keys } from '../../services';
 import { useInnerText } from '../inner_text';
 
 type ImageSize = 's' | 'm' | 'l' | 'xl' | 'fullWidth' | 'original';
+type Floats = 'left' | 'right';
+type Margins = 's' | 'm' | 'l' | 'xl';
 
 const sizeToClassNameMap: { [size in ImageSize]: string } = {
   s: 'euiImage--small',
@@ -46,6 +48,18 @@ const sizeToClassNameMap: { [size in ImageSize]: string } = {
   xl: 'euiImage--xlarge',
   fullWidth: 'euiImage--fullWidth',
   original: '',
+};
+
+const marginToClassNameMap: { [margin in Margins]: string } = {
+  s: 'euiImage--marginSmall',
+  m: 'euiImage--marginMedium',
+  l: 'euiImage--marginLarge',
+  xl: 'euiImage--marginXlarge',
+};
+
+const floatToClassNameMap: { [float in Floats]: string } = {
+  left: 'euiImage--floatLeft',
+  right: 'euiImage--floatRight',
 };
 
 export const SIZES = Object.keys(sizeToClassNameMap);
@@ -87,6 +101,14 @@ interface EuiImageProps extends CommonProps, HTMLAttributes<HTMLImageElement> {
    * When set to `true` will make the image clickable to a larger version
    */
   allowFullScreen?: boolean;
+  /**
+   * Float the image to the left or right. Useful in large text blocks.
+   */
+  float?: Floats;
+  /**
+   * Margin around the image.
+   */
+  margin?: Margins;
 }
 
 export const EuiImage: FunctionComponent<EuiImageProps> = ({
@@ -99,6 +121,8 @@ export const EuiImage: FunctionComponent<EuiImageProps> = ({
   fullScreenIconColor = 'light',
   alt,
   style,
+  float,
+  margin,
   ...rest
 }) => {
   const [isFullScreenActive, setIsFullScreenActive] = useState(false);
@@ -127,6 +151,8 @@ export const EuiImage: FunctionComponent<EuiImageProps> = ({
       'euiImage--hasShadow': hasShadow,
       'euiImage--allowFullScreen': allowFullScreen,
     },
+    margin ? marginToClassNameMap[margin] : null,
+    float ? floatToClassNameMap[float] : null,
     className
   );